Install TA .so files in $(libdir)/TAs
This change also adds an automake hook to prune .la files created by
libtool from the TA destination directory. This also adds a local
uninstall target to ensure that the resulting .so files are removed
during an uninstall as libtool does not remove them in the absence of
the corresponding .la files.

The .la files contain portable encoding of static and shared library
names and dependencies. These are usually safe to remove on Linux and
other ELF platforms where the runtime linker loads indirect library
dependencies automatically. On certain platforms, dependencies of the
compile module are not encoded in the binary, so the .la file is needed
so that lt_dlopen will find and load the correct dependencies. If
porting Open-TEE to such platforms (e.g. MacOS) the .la files should be
left intact.

Change-Id: I5223143df86349f66e758db18a90f83ce2b84704
Signed-off-by: Thomas Nyman <thomas.nyman@aalto.fi>